refactor(inbox): split covers reconciliation into a mixin to satisfy structural limits
CI's structural check caught what my piped local run swallowed:
global_inbox_runtime.py over the 2000-line file limit and two functions
over 100 lines. Move the finalize-time reconciliation methods to a
CoversReconciliationMixin (same pattern as InboxAdmissionMixin), extract
the renotice row transaction block and the mark_covered registration.
No behavior change.
